requirements.txt
flask
numpy
pandas
gunicorn
google-generativeai
wikipedia
python-dotenv
groq
scikit-learn